html/browsers/browsing-the-web/navigating-across-documents/010-form-action-javascript-url-aborted-by-navigation.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#queue-a-task">
<link rel="help" href="https://html.spec.whatwg.org/multipage/webappapis.html#dom-manipulation-task-source">
<title>Link with onclick form submit to navigate "javascript: url" in nested browsing context, is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<!-- Everything seen in the action="" is necessary to prevent a blank page in Chrome -->
<form target="test" action="javascript:parent.postMessage(location.href, '*');"></form>
<a target="test" onclick="document.forms[0].submit();reports.push('onclick');" href="href.html">Test</a>
<script>
var test = async_test();
var reports = [];

window.onmessage = test.step_func(function(e) {
  reports.push(e.data);
  // "href" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "href") {
    assert_equals(reports[0], "click");
    assert_equals(reports[1], "onclick");
    assert_equals(reports[2], "href");
    assert_equals(reports.length, 3);
  } else {


  }
  test.done();
});

test.step(function() {
  reports.push("click");
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates an iframe, a form with `action="javascript:...", and
    an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to submit the form, which targets the iframe.
3. The form action "javascript: url" navigation is queued in a task.
4. Navigating the iframe to the anchor's `href` must abort the iframe
    "javascript: url" navigation.
5. If the semantics are implemented correctly, reports will be
    ["click", "onclick", "href"].
    (There will be no report from the "javascript: url" code, because it
    is aborted by the navigation to "href.html")
-->
